Add callback in fetchObjectsOfType for error case
Problem: In fetchObjectsOfType in OddContentStore, if there is an error fetching an object, there is simply a return statement. The error never gets sent to the callback and thus users of this function will never be notified of an error. For example, in the objectsOfType function, fetchObjectsOfType is called, but if there is an error then the call back of objectsOfType will not be run. The result is that users of this code will never get called back and their code will spin indefinitely waiting.
Solution: add the callback that includes the error.